requirements.txt
asgiref==3.3.1
Django==3.0.13
django-phonenumber-field==5.0.0
phonenumbers==8.12.19
pytz==2021.1
sqlparse==0.4.1
